make group check in spyre conditional
